Refactor Artifacts (#22725)
* Use class method for`moved` artifact english name.

* Refactor linked artifacts.

* Refactor moved artifact.
