+++ lib/hbc/artifact/moved.rb
@@ -1,19 +1,13 @@
-class Hbc::Artifact::Moved < Hbc::Artifact::Base
-  ALT_NAME_ATTRIBUTE = "com.apple.metadata:kMDItemAlternateNames".freeze
+require "hbc/artifact/relocated"
 
-  attr_reader :source, :target
-
-  def english_name
-    self.class.artifact_english_name
-  end
-
+class Hbc::Artifact::Moved < Hbc::Artifact::Relocated
   def summary
     contents = @cask.artifacts[self.class.artifact_dsl_key].map do |artifact|
       summarize_artifact(artifact)
     end - [nil]
 
     {
-      english_description: "#{english_name}s managed by brew-cask:",
+      english_description: "#{self.class.artifact_english_name}s managed by brew-cask:",
       contents:            contents,
     }
   end
@@ -45,22 +39,10 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
   end
 
   def move
-    ohai "Moving #{english_name} '#{source.basename}' to '#{target}'"
+    ohai "Moving #{self.class.artifact_english_name} '#{source.basename}' to '#{target}'"
     target.dirname.mkpath
     FileUtils.move(source, target)
-  end
-
-  def load_specification(artifact_spec)
-    source_string, target_hash = artifact_spec
-    raise Hbc::CaskInvalidError if source_string.nil?
-    @source = @cask.staged_path.join(source_string)
-    if target_hash
-      raise Hbc::CaskInvalidError unless target_hash.respond_to?(:keys)
-      target_hash.assert_valid_keys(:target)
-      @target = Hbc.send(self.class.artifact_dirmethod).join(target_hash[:target])
-    else
-      @target = Hbc.send(self.class.artifact_dirmethod).join(source.basename)
-    end
+    add_altname_metadata target, source.basename.to_s
   end
 
   def preflight_checks
@@ -73,7 +55,7 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
       end
     end
     unless source.exist?
-      message = "It seems the #{english_name} source is not there: '#{source}'"
+      message = "It seems the #{self.class.artifact_english_name} source is not there: '#{source}'"
       raise Hbc::CaskError, message
     end
     true
@@ -81,16 +63,16 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
 
   def warning_target_exists
     message_parts = [
-                      "It seems there is already #{self.class.artifact_english_article} #{english_name} at '#{target}'",
+                      "It seems there is already #{self.class.artifact_english_article} #{self.class.artifact_english_name} at '#{target}'",
                     ]
     yield(message_parts) if block_given?
     message_parts.join("; ")
   end
 
   def delete
-    ohai "Removing #{english_name}: '#{target}'"
+    ohai "Removing #{self.class.artifact_english_name}: '#{target}'"
     if Hbc::MacOS.undeletable?(target)
-      raise Hbc::CaskError, "Cannot remove undeletable #{english_name}"
+      raise Hbc::CaskError, "Cannot remove undeletable #{self.class.artifact_english_name}"
     elsif force
       Hbc::Utils.gain_permissions_remove(target, command: @command)
     else
@@ -104,7 +86,7 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
     printable_target.sub!(%r{^'#{ENV['HOME']}/*}, "~/'")
 
     unless target.exist?
-      warning = "Missing #{english_name}"
+      warning = "Missing #{self.class.artifact_english_name}"
       warning = "#{Tty.red.underline}#{warning}#{Tty.reset}: "
     end
